feat: Added functionality to handle user number input
When a button is clicked, the event.target.dataset.value property retrieves the value of the clicked button.

Then, the code checks whether the userNum1 variable is null. If userNum1 is null, the code sets userNum1 to the clicked number, and sets the textContent of the display element to the value of userNum1.

If userNum1 is not null, the code concatenates the clicked number to the existing value of userNum1, and updates the textContent of the display element.