fix: display not updating on 2nd num btn click
The display was not updating correctly when a number button was clicked for the second time.

This was because the code was not checking whether userNum2 was null or not, and was updating the display with the wrong value.

To fix this issue, the code has been updated to check userNum2, and then update the display with the appropriate value.

The numberButtons event listener has been amended to check for userNum2 BEFORE updating the display, and to display the sum of userNum1 and userNum2 when userNum2 was not null.